online_inference/requirements.txt
fastapi~=0.78.0
pydantic~=1.9.1
starlette~=0.19.1
ml_in_prod_juliarty_ml_project_1~=0.2.0
gdown~=4.4.0
PyYAML~=6.0
scikit-learn~=1.0.2
pandas~=1.4.2
click~=8.1.3
requests~=2.27.1
uvicorn~=0.17.6